requirements.txt
openai>=1.0.0
anthropic>=0.18.1
google-generativeai>=0.3.0
scipy>=1.12.0
numpy>=1.26.0
matplotlib>=3.8.0
python-dotenv>=1.0.0 